requirements.txt
# Core dependencies for production
requests>=2.31.0
python-dotenv>=1.0.0
scipy>=1.11.0
psycopg2-binary>=2.9.9
